# authz-authn/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Lyfter JWT
  description: API specification generated from Flask blueprints for User, Product, Invoice, Order, and Contact services.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local development server
paths:
...
  /users/register:
    post:
      summary: Register a new user
      description: Creates a new user in the system. Requires admin privileges.
      tags:
        - Users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAuthRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          description: Missing or invalid required fields (username or password)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized or admin access required
        '409':
          description: Conflict (e.g., username already exists)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/login:
    post:
      summary: Authenticate user
      description: Logs in a user and returns access and refresh JWT tokens. Logs the attempt in login history.
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAuthRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          description: Invalid JSON, missing required fields, or empty values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid username or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server configuration error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me:
    get:
      summary: Get current authenticated user details
      tags:
        - Users
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Invalid or expired access token, or user no longer exists
...
  /users/login-history:
    get:
      summary: Retrieve login history
      description: Returns login attempts and audit records. Can be filtered by user ID. Requires admin privileges.
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: user_id
          required: false
          schema:
            type: integer
          description: Optional filter to retrieve login entries for a specific user ID
          example: 1
      responses:
        '200':
          description: List of login entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoginEntry'
        '400':
          description: Invalid user_id query parameter format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized or admin privilege required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==========================================
  # CONTACT ROUTES (/contacts)
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserAuthRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: john_doe
        password:
          type: string
          format: password
          example: securepassword123
    AuthTokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          description: Included if the endpoint issues refresh tokens (e.g., login, register).
          example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: john_doe
        role:
          type: string
          example: user
    LoginEntry:
      type: object
      properties:
        id:
          type: integer
          example: 101
        user_id:
          type: integer
          nullable: true
          description: The ID of the authenticated user, or null if login failed with an unknown username.
          example: 1
        ip_address:
          type: string
          example: "192.0.2.1"
        is_success:
          type: boolean
          example: true
        timestamp:
          type: string
          format: date-time
          example: "2026-08-23T21:30:00+00:00"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Something went wrong
...